Seperated out useful testing functions
