+++ src/app/reducers/tests/pollsState.test.ts
@@ -1,32 +1,11 @@
+import {
+  generateOptions,
+  generatePoll,
+  generatePolls
+} from "../../../tests/pollTestUtils";
 import * as types from "../../actions/action-types";
-import { Poll, PollOption } from "../../types";
+import { Poll } from "../../types";
 import reducer, { initialPollsState } from "../pollsState";
-
-const generateOptions: (n: number) => PollOption[] = n => {
-  const options: PollOption[] = [];
-  for (let i = 0; i < n; i++) {
-    options.push({
-      value: `value${i + 1}`,
-      optionId: `${i + 1}`,
-      votes: []
-    });
-  }
-  return options;
-};
-const generatePoll: (index: number) => Poll = i => ({
-  pollId: `${i}`,
-  pollName: `pollName${i}`,
-  options: generateOptions(3),
-  description: `description${i}`,
-  creator: { id: `${i}`, userName: `user${i}` }
-});
-const generatePolls: (n: number) => Poll[] = n => {
-  const polls = [];
-  for (let i = 0; i < n; i++) {
-    polls.push(generatePoll(i + 1));
-  }
-  return polls;
-};
 
 const loadingTest: (type: string) => void = type =>
   expect(reducer({ ...initialPollsState, error: "whoops" }, { type })).toEqual({
